Setup Jest and add health endpoint test
