+++ README.md
@@ -1,7 +1,7 @@
-
 # Web IRC Client – Backend
 
 ## Overview
+
 Node.js gateway that maps browser WebSocket traffic to IRC. **This project is intended to be run locally by a single user (you) on your own computer. It is not designed for deployment, public access, or multi-user scenarios.**
 
 One WebSocket connection may control multiple IRC sessions. Each client → server message must include a unique `id` so the backend knows which IRC connection it applies to. All server → client events include the same `id`.
@@ -9,6 +9,7 @@ One WebSocket connection may control multiple IRC sessions. Each client → serv
 Features related to authentication, production hardening, deployment, and multi-user support have been intentionally omitted or removed to keep the project simple for local use.
 
 ## Stack
+
 - Node.js 20
 - Express 5
 - irc‑framework
@@ -17,7 +18,9 @@ Features related to authentication, production hardening, deployment, and multi-
 ## Quick start
 
 ### Local
+
 Requires Node.js 20 or later.
+
 ```bash
 git clone <repo‑url>
 cd pulse-irc-backend
@@ -26,6 +29,7 @@ npm run dev
 ```
 
 ## Environment
+
 Copy `.env.example` to `.env` before running:
 
 ```bash
@@ -40,24 +44,34 @@ PORT=3000
 ```
 
 ## Scripts
+
 - `npm run dev` – dev with nodemon
 - `npm start` – prod
 - `npm test` – tests
 - `npm run lint` – ESLint
 - `npm run format` – Prettier
 
+## Testing
+
+Run all tests:
+
+```bash
+npm test
+```
+
 ## API
 
 ### WebSocket
+
 Endpoint: any path (example: `/ws`)
 
 Example messages:
+
 ```json
 { "type": "connect", "id": "server1", "server": "irc.example.net", "nick": "myNick" }
 { "type": "join", "id": "server1", "channel": "#chat" }
 { "type": "message", "id": "server1", "channel": "#chat", "text": "Hello" }
 ```
-
 
 <!--
 ### REST
@@ -65,7 +79,9 @@ Example messages:
 -->
 
 ## License
+
 MIT
 
 ## Contributing
+
 See [CONTRIBUTING.md](CONTRIBUTING.md) for development guidelines including linting and formatting.
