+++ package.json
@@ -4,31 +4,31 @@
 	"description": "Proxies WebSocket traffic to IRC for local single-user use.",
 	"main": "src/index.js",
 	"scripts": {
-		"test": "echo \"Error: no test specified\" && exit 1",
+		"test": "jest",
 		"lint": "eslint .",
 		"format": "prettier --write .",
 		"dev": "nodemon src/index.js",
-               "start": "node src/index.js"
-       },
-       "engines": {
-               "node": ">=20"
-       },
-       "author": "",
+		"start": "node src/index.js"
+	},
+	"engines": {
+		"node": ">=20"
+	},
+	"author": "",
 	"license": "MIT",
 	"devDependencies": {
 		"@eslint/js": "^9.31.0",
 		"eslint": "^9.31.0",
 		"globals": "^16.3.0",
 		"nodemon": "^3.1.10",
-		"prettier": "^3.6.2"
+		"prettier": "^3.6.2",
+		"jest": "^30.0.4",
+		"supertest": "^7.1.3"
 	},
 	"dependencies": {
 		"cors": "^2.8.5",
 		"dotenv": "^17.2.0",
 		"express": "^5.1.0",
 		"irc-framework": "^4.14.0",
-		"jest": "^30.0.4",
-		"supertest": "^7.1.3",
 		"ws": "^8.18.3"
 	}
 }
