+++ src/index.js
@@ -2,25 +2,9 @@
 require('dotenv').config();
 
 const http = require('http');
-const express = require('express');
-const cors = require('cors');
 const logger = require('./logger');
 const setupWebSocketServer = require('./wsHandler');
-
-const app = express();
-app.use(cors());
-
-app.get('/health', (req, res) => {
-	res.json({ status: 'ok' });
-});
-
-// Centralized error handler for REST endpoints
-/* eslint-disable no-unused-vars */
-app.use((err, req, res, next) => {
-	logger.error('Unhandled error:', err);
-	res.status(500).json({ error: 'Internal Server Error' });
-});
-/* eslint-enable no-unused-vars */
+const app = require('./app');
 
 const server = http.createServer(app);
 setupWebSocketServer(server);
